requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
einops>=0.6.1
timm>=0.9.0

# Data Processing
pandas>=2.0.0
xarray>=2023.1.0
netCDF4>=1.6.0
h5py>=3.8.0
scikit-image>=0.20.0

# ERA5 Data (Optional but recommended for real data)
cdsapi>=0.6.1  # Copernicus Climate Data Store API
cfgrib>=0.9.10  # GRIB format support (optional)

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.28.0

# Experiment Tracking
tensorboard>=2.12.0
wandb>=0.15.0

# Development/Testing
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

